Got ball position to reset to origin, finished scoring
I created a prefab for the invisible barrier that registers a score upon contact with the ball. I managed to get the ball position to reset to the origin/center of the screen/point (0, 0) when either player scores a point. I also got the scores to display correctly on the screen for each player when someone scores.